refactor: move asn1 parsing to utils folder to make accessible for multiple modules
