+++ lib/src/modules/pe/mod.rs
@@ -23,7 +23,6 @@ use crate::types::Struct;
 #[cfg(test)]
 mod tests;
 
-mod asn1;
 mod authenticode;
 pub mod parser;
 mod rva2off;
